+++ Source/RealismOverhaul.csproj
@@ -30,9 +30,9 @@
   <ItemGroup>
     <Compile Include="AdjustableCoMShifter.cs" />
     <Compile Include="ModuleROSolarPanel.cs" />
-    <Compile Include="DataTransmitterRO\AntennaType.cs" />
-    <Compile Include="DataTransmitterRO\ModuleDataTransmitterRO.cs" />
-    <Compile Include="DataTransmitterRO\TechLevel.cs" />
+    <Compile Include="Communications\AntennaType.cs" />
+    <Compile Include="Communications\ModuleAntennaRO.cs" />
+    <Compile Include="Communications\TechLevel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
